requirements.txt
boto3==1.26.94
fastapi==0.95.0
uvicorn==0.21.1
pandas~=1.5.3
psycopg2~=2.9.5
python-dotenv~=1.0.0
pyodbc
psycopg2-binary
pypyodbc~=1.3.5